backend/requirements.txt
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.1
Werkzeug==3.0.3
python-dotenv==1.0.1
requests==2.32.3
Pillow==10.4.0
pymysql==1.1.1

